Safari-611.1.10/LayoutTests/js/dom/modules/module-execution-error-should-be-propagated-to-onerror.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test window.onerror will be fired when the inlined module throws an error.');
// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
window.onerror = function () {
    finishJSTest();
}
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module">
debug('Executing the module.');
throw new Error("module is executed.");
</script>
</body>
</html>
