LayoutTests/js/dom/modules/module-execution-error-should-be-propagated-to-onerror-expected.txt

CONSOLE MESSAGE: line 19: Error: module is executed.
Test window.onerror will be fired when the inlined module throws an error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Executing the module.
PASS successfullyParsed is true

TEST COMPLETE
